fixed star rating iisue and , cover letter 11-07-2026

+++ src/migrations/20260416064250-create-exit-feedbacks.js
@@ -40,7 +40,7 @@ module.exports = {
       },
 
       rating: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(2, 1),
         allowNull: true,
       },
